requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
django-flatpickr==2.0.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
